requirements.txt
# Real-Time Meeting Intelligence Agent
# Core dependencies

# Audio processing
sounddevice>=0.5.0
soundfile>=0.12.0
scipy>=1.10.0
numpy>=1.24.0

# PDF parsing for RAG
pypdf>=4.0.0

# System monitoring
psutil>=5.9.0

# LLM inference (for answer generation)
llama-cpp-python>=0.3.0
